Add currentType to ShaderEnv
This patch defines a new type ShaderType with constructors for Vertex
and Pixel, and adds a new record currentType to ShaderEnv which is now
used during name/component resolution to ensure that inputs are only
resolved in the case of Vertex and outputs are only resolved in the case
of Pixel. In addition, the built-in name `position` is now only resolved
in the case of Vertex.