Insert with flipped const function
The parse function builds the function map from right to left. The
default map insertion function replaces old values with new values if
the key already exists in the map, but since we want the final
definition of a function to override previous definitions, we use the
insertWith function with the flipped version of the const function to
reverse the default behaviour.