Tidy up codebase
This patch renames PolarCore and PolarSys to Core and Sys respectively.
In addition, action propagation is rewritten to allow actions to be
written to the 'output' portion of Sys to be run by Core. This replaces
the old action propagation whereby the 'result' portion of Sys was
simply a list of Core actions. The new method of propagating a Core
action from Sys is through writing a SysCoreAction value.

The patch also completely rewrites the logWrite function to be part of a
typeclass and implements instances for both Core and Sys, allowing the
function to be used in both layers identically.