Storage: Lookup by any key type (#11)
* Make InnerStorage its own data structure

Previously InnerStorage was a type alias to a tuple as tuples of size 2
already have an instance of AsEmpty. This patch implements an instance
for the new InnerStorage type through the `nearly` function, using
`defaultInnerStorage` as the empty value and `innerStorageNull` as the
null condition.

* Rename InnerStorage to VectorStorage

* Storage: Lookup by any key type

This patch replaces the name=>index hashmap for vector storages with a
hashmap from hash=>index nested within another hashmap by TypeRep.

Essentially this allows arbitrary values to be added as a lookup key for
any index in a vector storage as long as that key is Hashable. This
increases the number of lookups and insertions by 1 but all lookups and
insertions still take place in amortized constant time!

This patch closes issue #9